classifier/pyproject.toml
[project]
name = "classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "loguru>=0.7.3",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
]
